requirements.txt
streamlit
yfinance
vaderSentiment
python-dotenv
pandas
matplotlib
plotly
requests
google-generativeai
yahooquery
